Amazon/requirements.txt
pymongo
django
djongo
djangorestframework
